Change to `events` package for better compatibility
This moves away from using `EventTarget` directly, as that is not available in
various environments such as Safari 13.1. As a replacement, this changes to
`events` which emulates the Node.js API in browser.

This is a breaking change, so clients will need to adjust code that adds
listeners to call the new function names.

Fixes https://github.com/vector-im/element-web/issues/15493