Add from widget "completed" action
This allows widgets to signal that they are now complete, with
whatever it is they wanted to do. This could for example be
used to tell the client user that they are now safe to close
a modal widget, should there be such restrictions that the
modal should not be closed before the widget has completed.

This is different from requesting to close the widget, since
we might still have information visible to the user.