Extend url template variables with device_id (#78)
* Extend url template variables with device_id

I mentioned the need for this parameter in the related [MSC 3891 ](https://github.com/matrix-org/matrix-spec-proposals/pull/3819) which implements to device messages. They are relying on the device id to be available, but there is currently no way to access a device id from within a widget (Element Call does this but is cheating :P). Therefore [I propose to include the device id in the available parameters](https://github.com/matrix-org/matrix-spec-proposals/pull/3819#discussion_r1099833846).

Signed-off-by: Oliver Sand <oliver.sand@nordeck.net>

* Add the `matrix_` prefix to the parameter

Signed-off-by: Dominik Henneke <dominik.henneke@nordeck.net>

---------

Signed-off-by: Oliver Sand <oliver.sand@nordeck.net>
Signed-off-by: Dominik Henneke <dominik.henneke@nordeck.net>
Co-authored-by: Dominik Henneke <dominik.henneke@nordeck.net>