Don't make error data payload extensible
as this makes it too easy for drivers to put data in the wrong section.

Still define the payload type as an interface so that it can be
extended in a future version of the API.

Also don't use a subfield now that non-extensibility makes the format of
the details fields unambiguous.