Define widget driver as an interface
- Adds `IWidgetDriver` interface
- Renames `WidgetDriver` to `BaseWidgetDriver` with old name (`WidgetDriver`) supported but deprecated

The goal here is to mean that a client that is implementing a driver can more easily see at compile time what is changing between versions.

However, this would mean that we need to change the versioning policy as any addition to IWidgetDriver would be a breaking change (so major version bump). Whereas, at the moment we don't do a major version bump because the abstract WidgetDriver class doesn't require the implementing client to implement anything.